Separate user_choice method to another file
